refactor uniform public views as one component
