+++ src/App.js
@@ -5,11 +5,8 @@ import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import { Container } from "reactstrap";
 import { AppProvider } from "./utilities/AppContext";
 import { AuthProvider, useAuth } from "./utilities/AuthContext";
-import LogIn from "./components/LogIn";
-import SignUp from "./components/SignUp";
+import Public from "./components/Public";
 import Profile from "./components/Profile";
-import LogOut from "./components/LogOut";
-import About from "./components/About";
 
 function App() {
   const auth = useAuth();
@@ -24,20 +21,18 @@ function App() {
                   {
                     auth.token
                       ? <Profile />
-                      : <LogIn />
+                      : <Public />
                   }
                 </Route>
-                <Route path="/signup">
-                  <SignUp />
-                </Route>
-                <Route path="/about">
-                  <About />
+                <Route path="/public/:view">
+                  {
+                    auth.token
+                      ? <Profile />
+                      : <Public />
+                  }
                 </Route>
                 <Route path={`/users/:userId`}>
                   <Profile />
-                </Route>
-                <Route path="/logout">
-                  <LogOut />
                 </Route>
               </Switch>
             </Router>
